--- src/livro/BuscarLivrosView.js
@@ -1,7 +1,0 @@
-export function BuscarLivrosView() {
-  return (
-    <div>
-      <h1>Teste</h1>
-    </div>
-  );
-}
