+++ src/paginas/BuscarLivrosView.js
@@ -1,0 +1,11 @@
+import { Header } from "./Header";
+
+export function BuscarLivrosView() {
+  return (
+    <main>
+      <Header />
+      <h1>Buscar Livros</h1>
+      {/* Conteúdo da página de busca */}
+    </main>
+  );
+}
